+++ infra/exception/dialect/type/mysql/src/main/java/org/apache/shardingsphere/infra/exception/mysql/mapper/MySQLDialectExceptionMapper.java
@@ -30,7 +30,7 @@ import org.apache.shardingsphere.infra.exception.dialect.exception.syntax.table.
 import org.apache.shardingsphere.infra.exception.dialect.exception.syntax.table.TableExistsException;
 import org.apache.shardingsphere.infra.exception.dialect.exception.transaction.TableModifyInTransactionException;
 import org.apache.shardingsphere.infra.exception.dialect.mapper.SQLDialectExceptionMapper;
-import org.apache.shardingsphere.infra.exception.mysql.exception.AccessDeniedException;
+import org.apache.shardingsphere.infra.exception.dialect.exception.connection.AccessDeniedException;
 import org.apache.shardingsphere.infra.exception.mysql.exception.DatabaseAccessDeniedException;
 import org.apache.shardingsphere.infra.exception.mysql.exception.ErrorGlobalVariableException;
 import org.apache.shardingsphere.infra.exception.mysql.exception.ErrorLocalVariableException;
